+++ src/public/Disconnect-odsc.ps1
@@ -2,15 +2,13 @@ function Disconnect-odsc {
     [CmdletBinding()]
     param()
 
-    begin {
-
-    }
-
     process {
         $script:ODSToken = $null
+        $script:ODSCloudEnvironment = 'Global'
+        $script:ODSGraphEndpoint = 'https://graph.microsoft.com'
     }
 
     end {
-        Write-Host "Disconnected."
+        Write-Host 'Disconnected.'
     }
 }